requirements.txt
aiohttp==3.6.2
# Editable Git install with no remote (alpha-vantage==2.2.0)
-e c:\users\leegi\pycharmprojects\botstock\venv\lib\site-packages
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
confuse==1.1.0
coverage==5.1
idna==2.9
JSON-Datetime==0.0.5
multidict==4.7.6
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-mock==1.8.0
sendgrid==6.3.1
simplejson==3.17.0
six==1.14.0
urllib3==1.25.9
yarl==1.4.2
